backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
sqlalchemy==2.0.32
psycopg>=3.2.0
alembic==1.13.2
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pytest==8.3.2
python-docx==1.1.2
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
